// packages/omo-central-webnative/src/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type AddCirclesTokenInput = {
  address: Scalars['String'];
  owner?: Maybe<AddCirclesWalletInput>;
  createdAt: Scalars['String'];
  createdInBlockNo: Scalars['Int'];
  createdInBlockHash: Scalars['String'];
};

export type CirclesWallet = {
  __typename?: 'CirclesWallet';
  address: Scalars['String'];
  ownToken?: Maybe<CirclesToken>;
  tokens?: Maybe<Array<CirclesToken>>;
  transfers?: Maybe<Array<CirclesTokenTransfer>>;
  trustRelations?: Maybe<Array<CirclesTrustRelation>>;
};

export type AddCirclesWalletInput = {
  address: Scalars['String'];
  ownToken?: Maybe<AddCirclesTokenInput>;
};
// ...
export type MutationAddCirclesTokenArgs = {
  data: AddCirclesTokenInput;
};